+++ activesupport/lib/active_support/current_attributes.rb
@@ -102,7 +102,14 @@ module ActiveSupport
       end
 
       # Declares one or more attributes that will be given both class and instance accessor methods.
-      def attribute(*names)
+      #
+      # ==== Options
+      #
+      # * <tt>:default</tt> - The default value for the attributes. If the value
+      # is a proc or lambda, it will be called whenever an instance is
+      # constructed. Otherwise, the value will be duplicated with +#dup+.
+      # Default values are re-assigned when the attributes are reset.
+      def attribute(*names, default: nil)
         invalid_attribute_names = names.map(&:to_sym) & INVALID_ATTRIBUTE_NAMES
         if invalid_attribute_names.any?
           raise ArgumentError, "Restricted attribute names: #{invalid_attribute_names.join(", ")}"
@@ -126,6 +133,8 @@ module ActiveSupport
         end
 
         singleton_class.delegate(*names.flat_map { |name| [name, "#{name}="] }, to: :instance, as: self)
+
+        defaults.merge! names.index_with { default }
       end
 
       # Calls this callback before #reset is called on the instance. Used for resetting external collaborators that depend on current values.
@@ -177,10 +186,12 @@ module ActiveSupport
         end
     end
 
+    class_attribute :defaults, instance_writer: false, default: {}
+
     attr_accessor :attributes
 
     def initialize
-      @attributes = {}
+      @attributes = merge_defaults!({})
     end
 
     # Expose one or more attributes within a block. Old values are returned after the block concludes.
@@ -200,8 +211,21 @@ module ActiveSupport
     # Reset all attributes. Should be called before and after actions, when used as a per-request singleton.
     def reset
       run_callbacks :reset do
-        self.attributes = {}
+        self.attributes = merge_defaults!({})
       end
     end
+
+    private
+      def merge_defaults!(attributes)
+        defaults.each_with_object(attributes) do |(name, default), values|
+          value =
+            case default
+            when Proc then default.call
+            else default.dup
+            end
+
+          values[name] = value
+        end
+      end
   end
 end